Fix imports after tool relocation
